list quanto fees, remove dexs as api is error-prone, as doesn't return timestamp on given day
